+++ ProjetoA3_CadastroDeAlunos/App.config
@@ -1,0 +1,8 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <connectionStrings>
+    <add name="MySqlConnection"
+     connectionString="Server=localhost;Database=teste;Uid=root;Pwd=admin;"
+     providerName="MySql.Data.MySqlClient"/>
+  </connectionStrings>
+</configuration>
