+++ ProjetoA3_CadastroDeAlunos/Src/Repositories/Connection.cs
@@ -1,0 +1,19 @@
+﻿using MySql.Data.MySqlClient;
+using System;
+using System.Collections.Generic;
+using System.Configuration;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ProjetoA3_CadastroDeAlunos.Src.Repositories
+{
+    internal class Connection
+    {
+        public static MySqlConnection GetConnection()
+        {
+            string connectionString = ConfigurationManager.ConnectionStrings["MySqlConnection"].ConnectionString;
+            return new MySqlConnection(connectionString);
+        }
+    }
+}
